backend/foodgram/requirements.txt
Django==3.2.15
django-filter==21.1
djangorestframework==3.11.2
djangorestframework-simplejwt==5.1.0
flake8==4.0.1
isort==5.10.1
gunicorn==20.0.4
PyJWT==2.1.0
requests==2.26.0
sqlparse==0.4.2
toml==0.10.2
pytz==2022.1
djoser==2.0.5
django-cors-headers==3.13.0
Pillow==9.2.0
psycopg2-binary==2.8.6
